requirements.txt
asgiref==3.8.1
channels==4.1.0
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.15.1
msgpack==1.0.8
sqlparse==0.4.4
tzdata==2024.1
